+++ CHANGELOG.md
@@ -156,6 +156,7 @@
   - Config selection/button control has been moved to the bottom above the status bar
 - Now logs all other QT messages to debug logging
   - Program will look for a .env file beside executable/start_ui.py that accepts **LOG_LEVEL** \*(mainly used for debugging QT errors in dev)\_
+- All console logs (when program is ran in **debug**) will be output to stdout
 
 ### Fixed